Add router for PHP's built-in webserver
The Apache configuration for vpn-user-portal makes use of aliases
to implement fancy URLs for some paths in the portal.
To make these fancy URLs work in the development setup (which makes
use of the PHP built-in webserver), a router script is required.
This change provides such a router script.

The router can be used by starting the development webserver like this:
```
$ cd vpn-user-portal
$ php -S localhost:8082 -t web dev/router.php
```